move board out
