+++ core/src/com/battleship/model/Board.kt
@@ -1,4 +1,4 @@
-package com.battleship.model.ui
+package com.battleship.model
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Color
